ui-refactor: fix import paths
