+++ packages/ui/src/component/form/textInput/TextInput.hook.ts
@@ -1,10 +1,4 @@
-import {
-  isGetState,
-  onMount,
-  Prop,
-  useEffect,
-  useRef,
-} from '@rvjs/core/reactive'
+import { isGetState, onMount, Prop, useEffect, useRef } from '@rvjs/core'
 
 const useTextInput = (disabled: Prop<boolean>) => {
   const inputRef = useRef<HTMLInputElement>()
