+++ packages/ui/src/component/layout/grid/Grid.ts
@@ -1,7 +1,14 @@
 import { boxSprinkles } from '@layout/box/Box.css.ts'
 import { GridStyleProps } from '@layout/grid/Grid.css.ts'
-import { Children, element, ElementType, StyleProps } from '@rvjs/core/dom'
-import { dynamic, Prop, RefObject } from '@rvjs/core/reactive'
+import {
+  Children,
+  dynamic,
+  element,
+  ElementType,
+  Prop,
+  RefObject,
+  StyleProps,
+} from '@rvjs/core'
 import { HTMLDivType } from '@type/element.ts'
 
 interface BoxProps extends GridStyleProps, HTMLDivType {
