+++ packages/ui/src/component/shell/header/headerGlobalBar/HeaderGlobalBar.ts
@@ -1,6 +1,5 @@
 import Flex from '@layout/flex/Flex.ts'
-import { Children } from '@rvjs/core/dom'
-import { prop } from '@rvjs/core/reactive'
+import { Children, prop } from '@rvjs/core'
 import { headerGlobalBar_style } from '@shell/header/headerGlobalBar/HeaderGlobalBar.css.ts'
 
 interface HeaderGlobalBarProps {
