+++ packages/ui/src/component/shell/header/headerHr/HeaderHr.ts
@@ -1,5 +1,5 @@
 import Box from '@layout/box/Box.ts'
-import { prop } from '@rvjs/core/reactive'
+import { prop } from '@rvjs/core'
 import { headerHr_style } from '@shell/header/headerHr/HeaderHr.css.ts'
 
 const HeaderHr = () => {
