+++ packages/ui/src/component/typography/codeSnippet/multi/MultiCodeSnippet.ts
@@ -3,8 +3,7 @@ import chevronDownSvg from '@icon/chevron--down.svg?element'
 import copySvg from '@icon/copy.svg?element'
 import Flex from '@layout/flex/Flex.ts'
 import Tooltip from '@overlay/tooltip/Tooltip.ts'
-import { code, pre, svg } from '@rvjs/core/dom'
-import { dynamic, prop, useState } from '@rvjs/core/reactive'
+import { code, dynamic, pre, prop, svg, useState } from '@rvjs/core'
 import { CodeSnippetProps } from '@typography/codeSnippet/CodeSnippet.props.ts'
 import { copyToClipboard } from '@typography/codeSnippet/CodeSnippet.util.ts'
 import {
