+++ packages/ui/src/component/typography/codeSnippet/single/SingleCodeSnippet.ts
@@ -2,8 +2,7 @@ import Button from '@form/button/Button.ts'
 import copySvg from '@icon/copy.svg?element'
 import Flex from '@layout/flex/Flex.ts'
 import Tooltip from '@overlay/tooltip/Tooltip.ts'
-import { span, svg } from '@rvjs/core/dom'
-import { dynamic, prop } from '@rvjs/core/reactive'
+import { dynamic, prop, span, svg } from '@rvjs/core'
 import { CodeSnippetProps } from '@typography/codeSnippet/CodeSnippet.props.ts'
 import { copyToClipboard } from '@typography/codeSnippet/CodeSnippet.util.ts'
 import {
