+++ packages/ui/src/type/element.ts
@@ -1,4 +1,4 @@
-import { ElementProps } from '@rvjs/core/dom'
+import { ElementProps } from '@rvjs/core'
 
 export type HTMLDivType = Partial<
   Omit<ElementProps<'div'>, 'align' | 'classes' | 'children' | 'style' | 'ref'>
